Move pypi.py to utils for cleaner imports
