+++ backend/api/_tests/test_pypi.py
@@ -3,7 +3,7 @@ from unittest.mock import patch
 
 from requests import HTTPError
 
-from api.pypi import query_pypi, get_plugin_pypi_metadata
+from utils.pypi import query_pypi, get_plugin_pypi_metadata
 from utils.test_utils import FakeResponse, plugin, plugin_list
