+++ backend/api/model.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from typing import Tuple, Dict, List
 
 from utils.github import get_github_metadata
-from api.pypi import query_pypi, get_plugin_pypi_metadata
+from utils.pypi import query_pypi, get_plugin_pypi_metadata
 from api.s3 import get_cache, cache
 from utils.utils import render_description, send_alert
 from api.zulip import notify_new_packages
