+++ backend/preview/preview.py
@@ -10,7 +10,7 @@ import os
 import json
 
 from utils.github import github_pattern, get_github_metadata, get_github_repo_url
-from api.pypi import get_plugin_pypi_metadata
+from utils.pypi import get_plugin_pypi_metadata
 
 def get_plugin_preview(repo_pth: str, dest_dir: str, is_local: bool = False) -> dict:
     """Get plugin preview metadata of package at repo_pth.
